librdmacm: Fix select function
Without the user_fds mapping, select() could set bits for internal fds
instead of the user fds the application passed in, so the wrong (or no)
sockets were reported as ready.

Keep the list of the fds that are sent to poll in order to know
which fd belongs to each rfd when returning the revents to the
fds list.

Signed-off-by: Batsheva Black <bblack@nvidia.com>